OFFICIAL/requirements.txt
streamlit
pandas
numpy
matplotlib
matplotlib.pyplot
plotly
plotly.express

requests
os
gwosc
gwosc.api
gwpy.timeseries 
gwosc.locate 

copy
base64

matplotlib
random
math
scipy
collections

sklearn.linear_model
sqlite3
itertools